move utility classes for file conversion into convert/utils
